add infer code
