+++ egs/libritts/TTS/local/compute_neural_codec_and_prepare_text_tokens.py
@@ -1,575 +1,1 @@
-#!/usr/bin/env python3
-# Copyright    2023                            (authors: Feiteng Li)
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-"""
-Phonemize Text and EnCodec Audio.
-
-Usage example:
-    python3 bin/tokenizer.py \
-        --src_dir ./data/manifests --output_dir ./data/tokenized
-
-"""
-import argparse
-import logging
-import os
-from pathlib import Path
-
-import torch
-import torch.multiprocessing
-from icefall.utils import get_executor
-from lhotse import CutSet, NumpyHdf5Writer
-from lhotse.recipes.utils import read_manifests_if_cached
-from tqdm.auto import tqdm
-
-from valle.data import (
-    AudioTokenConfig,
-    AudioTokenExtractor,
-    TextTokenizer,
-    tokenize_text,
-)
-# from valle.data.fbank import get_fbank_extractor
-from valle.utils import SymbolTable
-
-os.environ["PROTOCOL_BUFFERS_PYTHON_IMPLEMENTATION"] = "python"
-
-
-# Torch's multithreaded behavior needs to be disabled or
-# it wastes a lot of CPU and slow things down.
-# Do this outside of main() in case it needs to take effect
-# even when we are not invoking the main (e.g. when spawning subprocesses).
-torch.set_num_threads(1)
-torch.set_num_interop_threads(1)
-torch.multiprocessing.set_sharing_strategy("file_system")
-
-
-def get_args():
-    parser = argparse.ArgumentParser()
-
-    parser.add_argument(
-        "--src-dir",
-        type=Path,
-        default=Path("data/manifests"),
-        help="Path to the manifest files",
-    )
-    parser.add_argument(
-        "--output-dir",
-        type=Path,
-        default=Path("data/tokenized"),
-        help="Path to the tokenized files",
-    )
-    parser.add_argument(
-        "--text-extractor",
-        type=str,
-        default="espeak",
-        help="espeak or pypinyin or pypinyin_initials_finals",
-    )
-    parser.add_argument(
-        "--audio-extractor",
-        type=str,
-        default="Encodec",
-        help="Encodec or Fbank",
-    )
-    parser.add_argument(
-        "--dataset-parts",
-        type=str,
-        default="dev-clean test-clean",
-        help="Space separated dataset parts",
-    )
-    parser.add_argument(
-        "--prefix",
-        type=str,
-        default="libritts",
-        help="prefix of the manifest file",
-    )
-    parser.add_argument(
-        "--suffix",
-        type=str,
-        default="jsonl.gz",
-        help="suffix of the manifest file",
-    )
-    parser.add_argument(
-        "--batch-duration",
-        type=float,
-        default=400.0,
-        help="The maximum number of audio seconds in a batch."
-        "Determines batch size dynamically.",
-    )
-    parser.add_argument(
-        "--split",
-        type=int,
-        default=1,
-        help="Split the cut_set into multiple parts",
-    )
- 
-    return parser.parse_args()
-
-class PypinyinBackend:
-    """PypinyinBackend for Chinese. Most codes is referenced from espnet.
-    There are two types pinyin or initials_finals, one is
-    just like "ni1 hao3", the other is like "n i1 h ao3".
-    """
-
-    def __init__(
-        self,
-        backend="initials_finals",
-        punctuation_marks: Union[str, Pattern] = Punctuation.default_marks(),
-    ) -> None:
-        self.backend = backend
-        self.punctuation_marks = punctuation_marks
-
-    def phonemize(
-        self, text: List[str], separator: Separator, strip=True, njobs=1
-    ) -> List[str]:
-        assert isinstance(text, List)
-        phonemized = []
-        for _text in text:
-            _text = re.sub(" +", " ", _text.strip())
-            _text = _text.replace(" ", separator.word)
-            phones = []
-            if self.backend == "pypinyin":
-                for n, py in enumerate(
-                    pinyin(
-                        _text, style=Style.TONE3, neutral_tone_with_five=True
-                    )
-                ):
-                    if all([c in self.punctuation_marks for c in py[0]]):
-                        if len(phones):
-                            assert phones[-1] == separator.syllable
-                            phones.pop(-1)
-
-                        phones.extend(list(py[0]))
-                    else:
-                        phones.extend([py[0], separator.syllable])
-            elif self.backend == "pypinyin_initials_finals":
-                for n, py in enumerate(
-                    pinyin(
-                        _text, style=Style.TONE3, neutral_tone_with_five=True
-                    )
-                ):
-                    if all([c in self.punctuation_marks for c in py[0]]):
-                        if len(phones):
-                            assert phones[-1] == separator.syllable
-                            phones.pop(-1)
-                        phones.extend(list(py[0]))
-                    else:
-                        if py[0][-1].isalnum():
-                            initial = get_initials(py[0], strict=False)
-                            if py[0][-1].isdigit():
-                                final = (
-                                    get_finals(py[0][:-1], strict=False)
-                                    + py[0][-1]
-                                )
-                            else:
-                                final = get_finals(py[0], strict=False)
-                            phones.extend(
-                                [
-                                    initial,
-                                    separator.phone,
-                                    final,
-                                    separator.syllable,
-                                ]
-                            )
-                        else:
-                            assert ValueError
-            else:
-                raise NotImplementedError
-            phonemized.append(
-                "".join(phones).rstrip(f"{separator.word}{separator.syllable}")
-            )
-        return phonemized
-
-
-class TextTokenizer:
-    """Phonemize Text."""
-
-    def __init__(
-        self,
-        language="en-us",
-        backend="espeak",
-        separator=Separator(word="_", syllable="-", phone="|"),
-        preserve_punctuation=True,
-        punctuation_marks: Union[str, Pattern] = Punctuation.default_marks(),
-        with_stress: bool = False,
-        tie: Union[bool, str] = False,
-        language_switch: LanguageSwitch = "keep-flags",
-        words_mismatch: WordMismatch = "ignore",
-    ) -> None:
-        if backend == "espeak":
-            phonemizer = EspeakBackend(
-                language,
-                punctuation_marks=punctuation_marks,
-                preserve_punctuation=preserve_punctuation,
-                with_stress=with_stress,
-                tie=tie,
-                language_switch=language_switch,
-                words_mismatch=words_mismatch,
-            )
-        elif backend in ["pypinyin", "pypinyin_initials_finals"]:
-            phonemizer = PypinyinBackend(
-                backend=backend,
-                punctuation_marks=punctuation_marks + separator.word,
-            )
-        else:
-            raise NotImplementedError(f"{backend}")
-
-        self.backend = phonemizer
-        self.separator = separator
-
-    def to_list(self, phonemized: str) -> List[str]:
-        fields = []
-        for word in phonemized.split(self.separator.word):
-            # "ɐ    m|iː|n?"    ɹ|ɪ|z|ɜː|v; h|ɪ|z.
-            pp = re.findall(r"\w+|[^\w\s]", word, re.UNICODE)
-            fields.extend(
-                [p for p in pp if p != self.separator.phone]
-                + [self.separator.word]
-            )
-        assert len("".join(fields[:-1])) == len(phonemized) - phonemized.count(
-            self.separator.phone
-        )
-        return fields[:-1]
-
-    def __call__(self, text, strip=True) -> List[List[str]]:
-        if isinstance(text, str):
-            text = [text]
-
-        phonemized = self.backend.phonemize(
-            text, separator=self.separator, strip=strip, njobs=1
-        )
-        return [self.to_list(p) for p in phonemized]
-
-
-def tokenize_text(tokenizer: TextTokenizer, text: str) -> List[str]:
-    phonemes = tokenizer([text.strip()])
-    return phonemes[0]  # k2symbols
-
-
-def remove_encodec_weight_norm(model):
-    from encodec.modules import SConv1d
-    from encodec.modules.seanet import SConvTranspose1d, SEANetResnetBlock
-    from torch.nn.utils import remove_weight_norm
-
-    encoder = model.encoder.model
-    for key in encoder._modules:
-        if isinstance(encoder._modules[key], SEANetResnetBlock):
-            remove_weight_norm(encoder._modules[key].shortcut.conv.conv)
-            block_modules = encoder._modules[key].block._modules
-            for skey in block_modules:
-                if isinstance(block_modules[skey], SConv1d):
-                    remove_weight_norm(block_modules[skey].conv.conv)
-        elif isinstance(encoder._modules[key], SConv1d):
-            remove_weight_norm(encoder._modules[key].conv.conv)
-
-    decoder = model.decoder.model
-    for key in decoder._modules:
-        if isinstance(decoder._modules[key], SEANetResnetBlock):
-            remove_weight_norm(decoder._modules[key].shortcut.conv.conv)
-            block_modules = decoder._modules[key].block._modules
-            for skey in block_modules:
-                if isinstance(block_modules[skey], SConv1d):
-                    remove_weight_norm(block_modules[skey].conv.conv)
-        elif isinstance(decoder._modules[key], SConvTranspose1d):
-            remove_weight_norm(decoder._modules[key].convtr.convtr)
-        elif isinstance(decoder._modules[key], SConv1d):
-            remove_weight_norm(decoder._modules[key].conv.conv)
-
-
-class AudioTokenizer:
-    """EnCodec audio."""
-
-    def __init__(
-        self,
-        device: Any = None,
-    ) -> None:
-        # Instantiate a pretrained EnCodec model
-        model = EncodecModel.encodec_model_24khz()
-        model.set_target_bandwidth(6.0)
-        remove_encodec_weight_norm(model)
-
-        if not device:
-            device = torch.device("cpu")
-            if torch.cuda.is_available():
-                device = torch.device("cuda:0")
-
-        self._device = device
-
-        self.codec = model.to(device)
-        self.sample_rate = model.sample_rate
-        self.channels = model.channels
-
-    @property
-    def device(self):
-        return self._device
-
-    def encode(self, wav: torch.Tensor) -> torch.Tensor:
-        return self.codec.encode(wav.to(self.device))
-
-    def decode(self, frames: torch.Tensor) -> torch.Tensor:
-        return self.codec.decode(frames)
-
-
-@dataclass
-class AudioTokenConfig:
-    frame_shift: Seconds = 320.0 / 24000
-    num_quantizers: int = 8
-
-    def to_dict(self) -> Dict[str, Any]:
-        return asdict(self)
-
-    @staticmethod
-    def from_dict(data: Dict[str, Any]) -> "AudioTokenConfig":
-        return AudioTokenConfig(**data)
-
-class AudioTokenExtractor(FeatureExtractor):
-    name = "encodec"
-    config_type = AudioTokenConfig
-
-    def __init__(self, config: Optional[Any] = None):
-        super(AudioTokenExtractor, self).__init__(config)
-        self.tokenizer = AudioTokenizer()
-
-    def extract(
-        self, samples: Union[np.ndarray, torch.Tensor], sampling_rate: int
-    ) -> np.ndarray:
-        if not isinstance(samples, torch.Tensor):
-            samples = torch.from_numpy(samples)
-        if sampling_rate != self.tokenizer.sample_rate:
-            samples = convert_audio(
-                samples,
-                sampling_rate,
-                self.tokenizer.sample_rate,
-                self.tokenizer.channels,
-            )
-        if len(samples.shape) == 2:
-            samples = samples.unsqueeze(0)
-        else:
-            raise ValueError()
-
-        device = self.tokenizer.device
-        encoded_frames = self.tokenizer.encode(samples.detach().to(device))
-        codes = encoded_frames[0][0]  # [B, n_q, T]
-        if True:
-            duration = round(samples.shape[-1] / sampling_rate, ndigits=12)
-            expected_num_frames = compute_num_frames(
-                duration=duration,
-                frame_shift=self.frame_shift,
-                sampling_rate=sampling_rate,
-            )
-            assert abs(codes.shape[-1] - expected_num_frames) <= 1
-            codes = codes[..., :expected_num_frames]
-        return codes.cpu().squeeze(0).permute(1, 0).numpy()
-
-    @property
-    def frame_shift(self) -> Seconds:
-        return self.config.frame_shift
-
-    def feature_dim(self, sampling_rate: int) -> int:
-        return self.config.num_quantizers
-
-    def pad_tensor_list(self, tensor_list, device, padding_value=0):
-        # 计算每个张量的长度
-        lengths = [tensor.shape[0] for tensor in tensor_list]
-        # 使用pad_sequence函数进行填充
-        tensor_list = [torch.Tensor(t).to(device) for t in tensor_list]
-        padded_tensor = torch.nn.utils.rnn.pad_sequence(
-            tensor_list, batch_first=True, padding_value=padding_value
-        )
-        return padded_tensor, lengths
-
-    def extract_batch(self, samples, sampling_rate, lengths) -> np.ndarray:
-        samples = [wav.squeeze() for wav in samples]
-        device = self.tokenizer.device
-        samples, lengths = self.pad_tensor_list(samples, device)
-        samples = samples.unsqueeze(1)
-
-        if not isinstance(samples, torch.Tensor):
-            samples = torch.from_numpy(samples)
-        if len(samples.shape) != 3:
-            raise ValueError()
-        if sampling_rate != self.tokenizer.sample_rate:
-            samples = [
-                convert_audio(
-                    wav,
-                    sampling_rate,
-                    self.tokenizer.sample_rate,
-                    self.tokenizer.channels,
-                )
-                for wav in samples
-            ]
-            samples = torch.stack(samples, 0) # convert samples from list to tensor
-        # Extract discrete codes from EnCodec
-        with torch.no_grad():
-            encoded_frames = self.tokenizer.encode(samples.detach().to(device))
-        encoded_frames = encoded_frames[0][0]  # [B, n_q, T]
-        batch_codes = []
-        for b, length in enumerate(lengths):
-            codes = encoded_frames[b]
-            duration = round(length / sampling_rate, ndigits=12)
-            expected_num_frames = compute_num_frames(
-                duration=duration,
-                frame_shift=self.frame_shift,
-                sampling_rate=sampling_rate,
-            )
-            batch_codes.append(codes[..., :expected_num_frames])
-        return [codes.cpu().permute(1, 0).numpy() for codes in batch_codes]
-
-def main():
-    args = get_args()
-
-    dataset_parts = args.dataset_parts.replace("--dataset-parts", "").strip()
-    if dataset_parts == "all":  # LibriTTS
-        dataset_parts = [
-            "dev-clean",
-            "dev-other",
-            "test-clean",
-            "test-other",
-            "train-clean-100",
-            "train-clean-360",
-            "train-other-500",
-        ]
-    else:
-        dataset_parts = dataset_parts.replace("-p", "").strip().split(" ")
-
-    assert len(dataset_parts) >= 1
-
-    manifests = read_manifests_if_cached(
-        dataset_parts=dataset_parts,
-        output_dir=args.src_dir,
-        prefix=args.prefix,
-        suffix=args.suffix,
-        types=["recordings", "supervisions", "cuts"],
-    )
-
-    text_tokenizer = None
-    if args.text_extractor:
-        text_tokenizer = TextTokenizer(backend=args.text_extractor)
-
-    audio_extractor = None
-    if args.audio_extractor:
-        if args.audio_extractor == "Encodec":
-            audio_extractor = AudioTokenExtractor(AudioTokenConfig())
-        else:
-            assert args.audio_extractor == "Fbank"
-            audio_extractor = get_fbank_extractor()
-
-    Path(args.output_dir).mkdir(parents=True, exist_ok=True)
-    unique_symbols = set()
-    num_jobs = min(32, os.cpu_count())
-    logging.info(f"dataset_parts: {dataset_parts} manifests {len(manifests)}")
-
-    prefix = args.prefix
-    if prefix and not prefix.endswith("_"):
-        prefix = f"{prefix}_"
-    with get_executor() as ex:
-        for partition, m in manifests.items():
-            logging.info(
-                f"Processing partition: {partition} CUDA: {torch.cuda.is_available()}"
-            )
-            try:
-                cut_set = CutSet.from_manifests(
-                    recordings=m["recordings"],
-                    supervisions=m["supervisions"],
-                )
-            except Exception:
-                cut_set = m["cuts"]
-
-            # Split cut_set if split > 1
-            split = 1
-            if args.split > 1:
-                cut_sets = cut_set.split(args.split)
-                split = args.split
-            else:
-                cut_sets = [cut_set]
-
-            for idx, part in enumerate(cut_sets):
-                # AudioTokenizer
-                if args.audio_extractor:
-                    if args.audio_extractor == "Encodec":
-                        storage_path = (
-                            f"{args.output_dir}/{args.prefix}_encodec_{partition}_{idx if split > 1 else ''}"
-                        )
-                    else:
-                        storage_path = (
-                            f"{args.output_dir}/{args.prefix}_fbank_{partition}_{idx if split > 1 else ''}"
-                        )
-
-                    if args.prefix.lower() in ["ljspeech", "aishell", "baker", "wenetspeech4tts"]:
-                        part = part.resample(24000)
-
-                    with torch.no_grad():
-                        if (
-                            torch.cuda.is_available()
-                            and args.audio_extractor == "Encodec"
-                        ):
-                            part = part.compute_and_store_features_batch(
-                                extractor=audio_extractor,
-                                storage_path=storage_path,
-                                num_workers=num_jobs,
-                                batch_duration=args.batch_duration,
-                                collate=False,
-                                overwrite=True,
-                                storage_type=NumpyHdf5Writer,
-                            )
-                        else:
-                            part = part.compute_and_store_features(
-                                extractor=audio_extractor,
-                                storage_path=storage_path,
-                                num_jobs=num_jobs if ex is None else 64,
-                                executor=ex,
-                                storage_type=NumpyHdf5Writer,
-                            )
-
-                # TextTokenizer
-                if args.text_extractor:
-                    for c in tqdm(part):
-                        if args.prefix == "baker" and args.text_extractor == "labeled_pinyin":
-                            phonemes = c.supervisions[0].custom["tokens"]["text"]
-                            unique_symbols.update(phonemes)
-                        else:
-                            if args.prefix == "ljspeech":
-                                text = c.supervisions[0].custom["normalized_text"]
-                                text = text.replace(""", '"').replace(""", '"')
-                                phonemes = tokenize_text(text_tokenizer, text=text)
-                            elif args.prefix in ["aishell", "aishell2", "wenetspeech4tts", "libritts"]:
-                                phonemes = tokenize_text(
-                                    text_tokenizer, text=c.supervisions[0].text
-                                )
-                                if c.supervisions[0].custom is None:
-                                    c.supervisions[0].custom = {}
-                            else:
-                                raise NotImplementedError(f"{args.prefix}")
-                            c.supervisions[0].custom["tokens"] = {"text": phonemes}
-                            unique_symbols.update(phonemes)
-
-                # Save each part with an index if split > 1
-                cuts_filename = f"{prefix}cuts_{partition}.{idx if split > 1 else ''}.{args.suffix}"
-                part.to_file(f"{args.output_dir}/{cuts_filename}")
-                logging.info(f"Saved {cuts_filename}")
-
-    if args.text_extractor:
-        unique_phonemes = SymbolTable()
-        for s in sorted(list(unique_symbols)):
-            unique_phonemes.add(s)
-        logging.info(f"{len(unique_symbols)} unique phonemes: {unique_symbols}")
-
-        unique_phonemes_file = f"{args.output_dir}/unique_text_tokens.k2symbols"
-        unique_phonemes.to_file(unique_phonemes_file)
-
-
-if __name__ == "__main__":
-    formatter = (
-        "%(asctime)s %(levelname)s [%(filename)s:%(lineno)d] %(message)s"
-    )
-    logging.basicConfig(format=formatter, level=logging.INFO)
-    main()
+../../../wenetspeech4tts/TTS/local/compute_neural_codec_and_prepare_text_tokens.py
