+++ egs/wenetspeech4tts/TTS/shared
@@ -1,0 +1,1 @@
+../../../icefall/shared/
