+++ egs/wenetspeech4tts/TTS/valle/tokenizer.py
@@ -3,8 +3,8 @@ from typing import List, Tuple
 
 import numpy as np
 import torch
-
 from k2 import SymbolTable
+
 
 class TextTokenCollater:
     """Collate list of text tokens
@@ -52,15 +52,10 @@ class TextTokenCollater:
         self.token2idx = {token: idx for idx, token in enumerate(unique_tokens)}
         self.idx2token = [token for token in unique_tokens]
 
-    def index(
-        self, tokens_list: List[str]
-    ) -> Tuple[torch.Tensor, torch.Tensor]:
+    def index(self, tokens_list: List[str]) -> Tuple[torch.Tensor, torch.Tensor]:
         seqs, seq_lens = [], []
         for tokens in tokens_list:
-            assert (
-                all([True if s in self.token2idx else False for s in tokens])
-                is True
-            )
+            assert all([True if s in self.token2idx else False for s in tokens]) is True
             seq = (
                 ([self.bos_symbol] if self.add_bos else [])
                 + list(tokens)
@@ -103,10 +98,7 @@ class TextTokenCollater:
         )
 
         tokens_lens = torch.IntTensor(
-            [
-                len(seq) + int(self.add_eos) + int(self.add_bos)
-                for seq in tokens_seqs
-            ]
+            [len(seq) + int(self.add_eos) + int(self.add_bos) for seq in tokens_seqs]
         )
 
         return tokens_batch, tokens_lens
@@ -115,7 +107,5 @@ class TextTokenCollater:
 def get_text_token_collater(text_tokens_file: str) -> TextTokenCollater:
     text_tokens_path = Path(text_tokens_file)
     unique_tokens = SymbolTable.from_file(text_tokens_path)
-    collater = TextTokenCollater(
-        unique_tokens.symbols, add_bos=True, add_eos=True
-    )
+    collater = TextTokenCollater(unique_tokens.symbols, add_bos=True, add_eos=True)
     return collater
